Replace src -> shared
