+++ tests/shared/src/main/scala/ldbc/tests/model/CountryLanguage.scala
@@ -8,9 +8,7 @@ package ldbc.tests.model
 
 import ldbc.dsl.*
 import ldbc.dsl.codec.Codec
-
 import ldbc.query.builder.Table
-
 import ldbc.schema.Table as SchemaTable
 
 case class CountryLanguage(
