+++ tests/shared/src/test/scala/ldbc/tests/CodecTest.scala
@@ -6,21 +6,15 @@
 
 package ldbc.tests
 
-import java.time.*
-
-import com.mysql.cj.jdbc.MysqlDataSource
-
 import cats.effect.*
-
+import com.mysql.cj.jdbc.MysqlDataSource
+import jdbc.connector.{ConnectionProvider as JdbcProvider, *}
+import ldbc.connector.{ConnectionProvider as LdbcProvider, *}
+import ldbc.dsl.*
+import ldbc.sql.*
 import munit.*
 
-import ldbc.sql.*
-
-import ldbc.dsl.*
-
-import ldbc.connector.{ ConnectionProvider as LdbcProvider, * }
-
-import jdbc.connector.{ ConnectionProvider as JdbcProvider, * }
+import java.time.*
 
 class LdbcCodecTest extends CodecTest:
