+++ tests/shared/src/test/scala/ldbc/tests/TableQuerySelectConnectionTest.scala
@@ -6,23 +6,15 @@
 
 package ldbc.tests
 
-import com.mysql.cj.jdbc.MysqlDataSource
-
 import cats.effect.*
-
-import munit.*
-
-import ldbc.sql.*
-
+import com.mysql.cj.jdbc.MysqlDataSource
+import jdbc.connector.{ConnectionProvider as JdbcProvider, *}
+import ldbc.connector.{ConnectionProvider as LdbcProvider, *}
 import ldbc.dsl.*
-
 import ldbc.query.builder.*
-
-import ldbc.connector.{ ConnectionProvider as LdbcProvider, * }
-
-import jdbc.connector.{ ConnectionProvider as JdbcProvider, * }
-
+import ldbc.sql.*
 import ldbc.tests.model.*
+import munit.*
 
 class LdbcTableQuerySelectConnectionTest extends TableQuerySelectConnectionTest:
